az-func-audio/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

openai~=1.35.1
python-dotenv==1.0.1
requests
azure-cosmos==4.8.0
reportlab==4.2.5
azure-functions
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
python-dotenv==1.0.1